fix: make default value an empty string instead of null
BREAKING CHANGE:
Before this change, the default value was null, which implied the user can simply add something, in fact, he had to include and escape quotes: something in order to pass down a string value to json.
After this change, the string values provided to the JSON are always encasulated in in quotes removing the need for double quoting. ui_banner_dismissible is treated as boolean and is _not_ quoted.